--- config/doctrine/Creator.orm.yml
@@ -1,40 +1,0 @@
-App\Entity\Creator:
-  type: entity
-  inheritanceType: SINGLE_TABLE
-  table: 'creator'
-  repositoryClass: App\Repository\CreatorRepository
-  discriminatorColumn:
-      name: discr
-      type: string
-  discriminatorMap:
-      creatorClient: CreatorClient
-      creatorUser: CreatorUser
-      creatorCenter: CreatorCentre
-  id:
-    id:
-      type: integer
-      id: true
-      generator:
-        strategy: AUTO
-      options:
-          unsigned: true
-  manyToOne:
-    document:
-      targetEntity: App\Entity\Attributes\Document
-      inversedBy: creators
-    user:
-      targetEntity: App\Entity\User
-      inversedBy: creators
-    note:
-      targetEntity: App\Entity\Attributes\Note
-      inversedBy: creators
-    evenement:
-      targetEntity: App\Entity\Attributes\Evenement
-      inversedBy: creators
-    contact:
-      targetEntity: App\Entity\Attributes\Contact
-      inversedBy: creators
-    dossier:
-      targetEntity: App\Entity\Attributes\Dossier
-      inversedBy: creators
-  lifecycleCallbacks: {  }
