+++ src/DataFixtures/FixturesPurger.php
@@ -12,12 +12,12 @@ use App\Entity\Attributes\Client;
 use App\Entity\Attributes\ClientCentre;
 use App\Entity\Attributes\ClientEntity;
 use App\Entity\Attributes\Contact;
+use App\Entity\Attributes\Creator;
 use App\Entity\Attributes\Document;
 use App\Entity\Attributes\Dossier;
 use App\Entity\Attributes\Evenement;
 use App\Entity\Attributes\MembreCentre;
 use App\Entity\Attributes\Note;
-use App\Entity\Creator;
 use App\Entity\Gestionnaire;
 use App\Entity\Membre;
 use App\Entity\RefreshToken;
