+++ src/Entity/Attributes/Document.php
@@ -13,7 +13,6 @@ use ApiPlatform\Metadata\Post;
 use App\Api\Filters\FolderIdFilter;
 use App\Controller\Api\UploadDocumentController;
 use App\Domain\Anonymization\AnonymizationHelper;
-use App\Entity\Creator;
 use App\Entity\Interface\FolderableEntityInterface;
 use App\Entity\SharedDocument;
 use App\Repository\DocumentRepository;
