+++ src/Entity/Attributes/DonneePersonnelle.php
@@ -2,7 +2,6 @@
 
 namespace App\Entity\Attributes;
 
-use App\Entity\Creator;
 use App\Entity\Traits\CreatorTrait;
 use App\Entity\User;
 use Doctrine\Common\Collections\ArrayCollection;
