+++ src/Entity/Membre.php
@@ -13,6 +13,7 @@ use App\Entity\Attributes\Beneficiaire;
 use App\Entity\Attributes\Centre;
 use App\Entity\Attributes\ClientMembre;
 use App\Entity\Attributes\ConsultationBeneficiaire;
+use App\Entity\Attributes\Creator;
 use App\Entity\Attributes\Evenement;
 use App\Entity\Attributes\MembreCentre;
 use App\Entity\Attributes\Subject;
