+++ src/Entity/User.php
@@ -21,6 +21,7 @@ use App\Entity\Attributes\Beneficiaire;
 use App\Entity\Attributes\BeneficiaireCentre;
 use App\Entity\Attributes\Centre;
 use App\Entity\Attributes\Client;
+use App\Entity\Attributes\Creator;
 use App\Entity\Attributes\MembreCentre;
 use App\Entity\Attributes\UserCentre;
 use App\Repository\UserRepository;
